requirements.txt
appdirs==1.4.4
asgiref==3.3.1
black==20.8b1
boto3==1.17.43
botocore==1.20.43
click==7.1.2
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.5
django-cors-headers==3.7.0
django-heroku==0.3.1
django-storages==1.11.1
django-taggit==1.3.0
djangorestframework==3.12.2
djangorestframework-camel-case==1.2.0
djangorestframework-simplejwt==4.6.0
filelock==3.0.12
gunicorn==20.1.0
jmespath==0.10.0
mypy-extensions==0.4.3
pathspec==0.8.1
Pillow==8.1.0
psycopg2==2.8.6
PyJWT==2.0.1
python-dateutil==2.8.1
pytz==2020.5
regex==2020.11.13
s3transfer==0.3.6
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.4
virtualenv==20.3.1
virtualenvwrapper-win==1.2.6
whitenoise==5.2.0
